Adding a length field to the String object
While chars.len isn't slow, using a dedicated length field offers a
performance optimization, especially in performance-critical sections
of code where string lengths are accessed often or where strings are large.
It's a micro-optimization, but one that can accumulate benefits in
large-scale applications. The primary benefit is improved predictability
and cache locality.

And also, it follows the book: "Crafting Interpreters".