Implement SET_LOCAL instruction for local variable assignments
The SET_LOCAL instruction take the slot number of the local
variable as its operand (i.e the slot number follows the
SET_LOCAL instruction in the Chunk.

It then stores the value at the specified slot in the stack.
Note that it does not pop the value on the stack.

Note also that it is up to the Compiler to keep track of
where local variables are stored on the stack and to emit
the proper slot operand values into the generated Chunk.